Add AsyncHandler to Routes

+++ backend/routes/userRouter.js
@@ -1,12 +1,13 @@
 let express = require("express");
 let body_parser = require("body-parser");
 let fetch = require("node-fetch");
+let asyncHandler = require("express-async-handler");
+let mongoose = require("mongoose");
 
 let User = require("../models/user");
 let Group = require("../models/group");
 let Request = require("../models/request");
 const session = require("express-session");
-const req = require("express/lib/request");
 
 let userRouter = express.Router();
 
@@ -14,18 +15,33 @@ userRouter.use(body_parser.urlencoded({ extended: false }));
 userRouter.use(body_parser.json());
 
 //Delete User =>Admin
-userRouter.delete("/deleteUser", async (req, res) => {
-  let userId = req.body.userId;
+userRouter.delete(
+  "/deleteUser",
+  asyncHandler(async (req, res) => {
+    let userId = req.body.userId;
+    console.log(userId);
 
-  let response = await User.deleteOne({ _id: userId });
-  res.send(JSON.stringify("user deleted successfully " + response));
-});
+    if (!mongoose.isValidObjectId(userId)) {
+      res.status(403);
+      throw new Error("userId Is not valid");
+    }
+
+    let response = await User.deleteOne({ _id: userId });
+    res.send(JSON.stringify(response));
+  })
+);
 
 //change handle
-userRouter.put("/changeHandle", async (req, res) => {
-  try {
+userRouter.put(
+  "/changeHandle",
+  asyncHandler(async (req, res) => {
     let userId = session.currentUser.id;
     let handle = req.body.handle;
+
+    if (!mongoose.isValidObjectId(userId)) {
+      res.status(403);
+      throw new Error("userId Is not valid");
+    }
 
     //chaeck if the new handle is existed on codeforces
     let result = await fetch(
@@ -41,9 +57,7 @@ userRouter.put("/changeHandle", async (req, res) => {
     user.handle = handle;
     let response = await result.save();
     res.send(JSON.stringify("handle has changed successfully " + response));
-  } catch (err) {
-    res.send(err.message);
-  }
-});
+  })
+);
 
 module.exports = userRouter;
